修改vim-docker安装指令

+++ b-vim/vim.config/install-vim-for-docker.sh
@@ -53,16 +53,19 @@ else if [ $SYS_VERSION = 'Linux' ];then
 	fi
 fi
 
+if [ -e $HOME/.vim/bundle/molokai/colors/molokai.vim ];then
 mv $HOME/.vim/bundle/molokai/colors/molokai.vim $HOME/.vim/bundle/molokai/colors/molokai.vim.bk
 cp $CURRENT_DIR/project_vimrc/molokai.vim $HOME/.vim/bundle/molokai/colors/molokai.vim
+fi
+if [ -e $HOME/.vim/bundle/vim-syntax-match/colors/syntaxmatch.vim ];then
+	mv $HOME/.vim/bundle/vim-syntax-match/colors/syntaxmatch.vim $HOME/.vim/bundle/vim-syntax-match/colors/syntaxmatch.vim.bk
+	cp $CURRENT_DIR/project_vimrc/syntaxmatch.vim  $HOME/.vim/bundle/vim-syntax-match/colors
+fi
+if [ -d /usr/share/vim/vim80/syntax ];then
+	sudo mv /usr/share/vim/vim80/syntax/c.vim /usr/share/vim/vim80/syntax/c.vim.bk
+	sudo cp $CURRENT_DIR/project_vimrc/c.vim /usr/share/vim/vim80/syntax/c.vim
 
-mv $HOME/.vim/bundle/vim-syntax-match/colors/syntaxmatch.vim $HOME/.vim/bundle/vim-syntax-match/colors/syntaxmatch.vim.bk
-cp $CURRENT_DIR/project_vimrc/syntaxmatch.vim  $HOME/.vim/bundle/vim-syntax-match/colors
-
-sudo mv /usr/share/vim/vim80/syntax/c.vim /usr/share/vim/vim80/syntax/c.vim.bk
-sudo cp mv $CURRENT_DIR/project_vimrc/c.vim /usr/share/vim/vim80/syntax/c.vim
-
-sudo mv /usr/share/vim/vim80/syntax/cpp.vim /usr/share/vim/vim80/syntax/cpp.vim.bk
-sudo cp mv $CURRENT_DIR/project_vimrc/cpp.vim /usr/share/vim/vim80/syntax/cpp.vim
-
+	sudo mv /usr/share/vim/vim80/syntax/cpp.vim /usr/share/vim/vim80/syntax/cpp.vim.bk
+	sudo cp $CURRENT_DIR/project_vimrc/cpp.vim /usr/share/vim/vim80/syntax/cpp.vim
+fi
